refactor(application): Gemini 재리뷰 피드백 반영 - 타입 안정성 개선 (PR#4)
## 변경 사항

### 1. TokenAssembler - 안전한 타입 변환
**파일**: `TokenAssembler.java:80`
**문제**: long → int 강제 형변환으로 오버플로우 위험
**해결**: `Math.toIntExact()` 사용으로 오버플로우 시 ArithmeticException 발생

**Before**:
```java
final int expiresInSeconds = (int) accessToken.remainingValidity().toSeconds();
```

**After**:
```java
final int expiresInSeconds = Math.toIntExact(accessToken.remainingValidity().toSeconds());
```

### 2. LoginService - Javadoc 예외 명세 정확화
**파일**: `LoginService.java:124-127`
**문제**: 실제 발생하는 예외와 문서 불일치
**해결**: 도메인 예외 명세 추가

**Before**:
```java
@throws IllegalStateException 사용자가 존재하지 않거나 ACTIVE 상태가 아닌 경우
```

**After**:
```java
@throws UserNotFoundException 사용자가 존재하지 않는 경우
@throws InvalidUserStatusException 사용자가 ACTIVE 상태가 아닌 경우
```

### 3. 아키텍처 결정: Command.credentialType 타입 유지 (SKIP)
**Gemini 제안**: `String credentialType` → `CredentialType` Enum 변경
**결정**: **제안 거부 (Skip)**

**이유**:
- ✅ **Application Layer Port 설계 원칙**: Port는 Domain에 의존하지 않음
- ✅ **헥사고날 아키텍처 경계**: Port는 Primitive 타입 사용
- ✅ **REST Adapter 바인딩**: `@RequestParam String` → Enum 변환은 Adapter 책임
- ✅ **현재 구조 유지**: Port(String) → Service(Enum 변환 + 검증)

## Gemini Code Assist 피드백 반영

### 처리된 이슈 (2건)
1. ✅ **MEDIUM: TokenAssembler 타입 안전성 개선**
   - Math.toIntExact() 사용으로 오버플로우 방지
   - ArithmeticException으로 명확한 오류 감지

2. ✅ **MEDIUM: LoginService Javadoc 정확성 향상**
   - 실제 예외와 문서 일치
   - 도메인 예외 명세로 계약 명확화

### Skip된 이슈 (1건)
- ⏭️ **MEDIUM: Command credentialType Enum 변환**
  - 아키텍처 원칙과 충돌
  - 현재 설계가 헥사고날 아키텍처에 부합

## 테스트 결과
- ✅ Total: 14 tests (TokenAssemblerTest + LoginServiceTest)
- ✅ All tests passed

+++ application/src/main/java/com/ryuqq/authhub/application/auth/assembler/TokenAssembler.java
@@ -76,7 +76,8 @@ public class TokenAssembler {
 
         // ✅ Token의 행위 메서드 활용 - remainingValidity()
         // Law of Demeter 준수 - Token이 직접 제공하는 메서드 사용
-        final int expiresInSeconds = (int) accessToken.remainingValidity().toSeconds();
+        // Math.toIntExact() 사용으로 오버플로우 시 ArithmeticException 발생 (안전한 변환)
+        final int expiresInSeconds = Math.toIntExact(accessToken.remainingValidity().toSeconds());
 
         return new LoginUseCase.Response(
                 accessTokenValue,
